+++ HISTORY.md
@@ -1,8 +1,8 @@
 # Rocksdb Change Log
 ## Unreleased
 
-### Bu Fixes
-* Shouldn't ignore return value of fsync() in flush.
+### New Features
+* RateLimiter support for throttling background reads, or throttling the sum of background reads and writes. This can give more predictable I/O usage when compaction reads more data than it writes, e.g., due to lots of deletions.
 
 ## 5.5.0 (05/17/2017)
 ### New Features